Update SharePipeline, fix share handling and notification generation

+++ app/Jobs/SharePipeline/UndoSharePipeline.php
@@ -33,35 +33,39 @@ class UndoSharePipeline implements ShouldQueue
 	{
 		$status = $this->status;
 		$actor = $status->profile;
-		$parent = $status->parent();
-		$target = $status->parent()->profile;
+		$parent = Status::find($status->reblog_of_id);
 
-		ReblogService::removePostReblog($parent->id, $status->id);
+		if($parent) {
+			$target = $parent->profile_id;
+			ReblogService::removePostReblog($parent->id, $status->id);
 
-		if ($status->uri !== null) {
+			if($parent->reblogs_count > 0) {
+				$parent->reblogs_count = $parent->reblogs_count - 1;
+				$parent->save();
+				StatusService::del($parent->id);
+			}
+
+			$notification = Notification::whereProfileId($target)
+				->whereActorId($status->profile_id)
+				->whereAction('share')
+				->whereItemId($status->reblog_of_id)
+				->whereItemType('App\Status')
+				->first();
+
+			if($notification) {
+				$notification->forceDelete();
+			}
+		}
+
+		if ($status->uri != null) {
 			return;
 		}
 
-		if($target->domain === null) {
-			Notification::whereProfileId($target->id)
-			->whereActorId($status->profile_id)
-			->whereAction('share')
-			->whereItemId($status->reblog_of_id)
-			->whereItemType('App\Status')
-			->delete();
+		if(config_cache('federation.activitypub.enabled') == false) {
+			return $status->delete();
+		} else {
+			return $this->remoteAnnounceDeliver();
 		}
-
-		$this->remoteAnnounceDeliver();
-
-		if($parent->reblogs_count > 0) {
-			$parent->reblogs_count = $parent->reblogs_count - 1;
-			$parent->save();
-			StatusService::del($parent->id);
-		}
-
-		$status->forceDelete();
-
-		return 1;
 	}
 
 	public function remoteAnnounceDeliver()
@@ -124,5 +128,8 @@ class UndoSharePipeline implements ShouldQueue
 
 		$promise->wait();
 
+		$status->delete();
+
+		return 1;
 	}
 }
